+++ app/src/main/java/by/zhenyabigel/bankingapplication/MainActivity.kt
@@ -4,14 +4,6 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import by.zhenyabigel.bankingapplication.screen.HomeScreen
 import by.zhenyabigel.bankingapplication.ui.theme.BankingApplicationTheme
 
 class MainActivity : ComponentActivity() {
@@ -19,25 +11,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            BankingApplicationTheme {
-               HomeScreen()
-            }
+            BankingApplicationTheme {}
         }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    BankingApplicationTheme {
-        Greeting("Android")
     }
 }
