+++ app/src/main/java/by/zhenyabigel/bankingapplication/model/Transaction.kt
@@ -1,8 +1,4 @@
 package by.zhenyabigel.bankingapplication.model
-
-import androidx.annotation.DrawableRes
-import java.util.UUID
-
 data class Transaction (
     val company: String,
     val transactionNumber: String,
