+++ app/src/main/java/by/zhenyabigel/bankingapplication/screen/home_screen/GetColorForStatus.kt
@@ -1,0 +1,15 @@
+package by.zhenyabigel.bankingapplication.screen.home_screen
+
+import androidx.compose.ui.graphics.Color
+import by.zhenyabigel.bankingapplication.ui.theme.Green
+import by.zhenyabigel.bankingapplication.ui.theme.Maroon
+import by.zhenyabigel.bankingapplication.ui.theme.Yellow
+
+fun getColorForStatus(status: String): Color {
+    return when (status) {
+        "Executed" -> Green
+        "In progress" -> Yellow
+        "Declined" -> Maroon
+        else -> Color.Gray // Цвет по умолчанию
+    }
+}
